Added upgrades suggested by Mark
"There have been a couple things brought up since the last software update for the Pi’s at each machine.  Some of this was due to some auditing that I did to see how accurate the operators were collecting data compared to the shiftlead reports.

The first thing that came ups was that when the operators view the last commits on a particular machine, all of the last commits show up even the ones they have from another machine.  It makes it difficult to look back and see which specific machines had what data collected.  I think this is difficult to explain in an email so if you need me to explain further try and give me a call.  

The second thing that was brought up was in line with this, but if there was a way on the user interface to have a counter that would show how many good and bad commits have been entered. "